+++ snyk_apiweb/__init__.py
@@ -1,6 +1,7 @@
-__version__ = "1.1.3"
+__version__ = "1.2.0"
 
 __all__ = [
+    "audit",
     "config",
     "probely_client",
     "tools",